requirements.txt
# Core Framework
Flask==3.0.0
Flask-SQLAlchemy==3.1.1
Flask-Migrate==4.0.5
Flask-Login==0.6.3
Werkzeug==3.0.1

# Database
psycopg2-binary==2.9.9
SQLAlchemy==2.0.23

# Task Queue
celery==5.3.4
redis==5.0.1
gevent==23.9.1
psycogreen==1.0.2

# Telegram
Telethon==1.33.1
cryptg==0.4.0
PySocks==1.7.1

# Data Processing
pandas==2.1.4
openpyxl==3.1.2

# HTTP & Requests
requests==2.31.0
aiohttp==3.9.1

# Utilities
python-dotenv==1.0.0
bcrypt==4.1.2
python-dateutil==2.8.2

# Validation
email-validator==2.1.0

# Reporting
reportlab==4.0.7

# Monitoring
psutil==5.9.6
cryptography==41.0.7
opentele==1.15.1
tgcrypto==1.2.5
